+++ app/controllers/application.js
@@ -1,9 +1,8 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
-  needs: ['currentUser'],
-  currentUser: Ember.computed.oneWay('controllers.currentUser'),
-  preferences: Ember.computed.oneWay('currentUser.preferences'),
+  currentUser: Ember.inject.service(),
+  preferences: Ember.inject.service(),
 
   navigationItems: [{
     id: 'dashboard', name: 'Dashboard', routeName: 'dashboard'
@@ -15,13 +14,27 @@ export default Ember.Controller.extend({
     id: 'performance', name: 'Performance', routeName: 'performance'
   }],
 
+  actionItems: [{
+    name: 'feedback',
+    label: 'Feedback'
+  }, {
+    name: 'logout',
+    label: 'Logout'
+  }],
+
   actions: {
     savePreferences: function() {
-      this.get('preferences').save();
+      this.get('preferences.model').save();
     },
 
     logout: function() {
-      window.location.replace(`${EmberENV.pistachioUrl}/logout`);
+      this.get('currentUser').logout();
+    },
+
+    feedback: function() {
+      /* jshint newcap: false */
+      Intercom('show');
+      /* jshint newcap: true */
     }
   }
 });