+++ app/controllers/dashboard/interaction.js
@@ -1,14 +1,23 @@
 import Ember from 'ember';
 import ModelsNavigationMixin from 'ember-cli-paint/mixins/models-navigation';
+import InteractionCompletion from 'phoenix/models/interaction-completion';
+import PromiseController from 'phoenix/controllers/promise';
 import notify from 'phoenix/helpers/notify';
 import InteractionCancellation from 'phoenix/services/interaction-cancellation';
-import PromiseController from 'phoenix/controllers/promise';
+
+var qualityOptionsMapping = {
+  'good': 'Good',
+  'bad': 'Bad'
+};
 
 export default Ember.Controller.extend(ModelsNavigationMixin, {
   needs: ['dashboard'],
   dashboard: Ember.computed.oneWay('controllers.dashboard'),
   navigableModels: Ember.computed.oneWay('dashboard.scheduledInteractions'),
   modelRouteParams: ['dashboard.interaction'],
+
+  showForm: false,
+  requestPromise: null,
 
   profiles: Ember.computed('model.advisor', 'model.clientContact', function() {
     return [{
@@ -25,6 +34,15 @@ export default Ember.Controller.extend(ModelsNavigationMixin, {
   checklistItems: Ember.computed.sort('model.checklistItems', 'checklistItemsSorting'),
   checklistItemsSorting: ['completed', 'createdAt'],
 
+  qualityOptions: Ember.computed(function() {
+    return InteractionCompletion.qualityOptions.map(function(option) {
+      return Ember.Object.create({
+        id: option,
+        name: qualityOptionsMapping[option]
+      });
+    });
+  }),
+
   _cancel: function(withdrawFromCompliance = false) {
     var requestPromise =
       InteractionCancellation.create().cancel(this.get('model'), response => {
@@ -38,8 +56,23 @@ export default Ember.Controller.extend(ModelsNavigationMixin, {
   },
 
   actions: {
+    chargeClient: function() {
+      this.set('requestPromise', PromiseController.create({
+        promise: this.get('completion').save().then(() => {
+          notify('The interaction has been completed.');
+          this.get('sidePanel').send('close');
+        }).catch(function() {
+          notify('There has been an error completing the interaction.', 'error');
+        })
+      }));
+    },
+
     hideSidePanel: function() {
       this.transitionToRoute('dashboard');
+    },
+
+    toggleForm: function() {
+      this.toggleProperty('showForm');
     },
 
     cancel: function() {
@@ -54,6 +87,7 @@ export default Ember.Controller.extend(ModelsNavigationMixin, {
       var model = this.get('model');
 
       model.set('scheduledCallTime', null);
+      model.set('actioned', false);
 
       this.set('requestPromise', PromiseController.create({
         promise: model.save().then(