+++ app/routes/dashboard/unused-advisor.js
@@ -2,13 +2,20 @@ import Ember from 'ember';
 import SidePanelRouteMixin from 'ember-cli-paint/mixins/side-panel-route';
 
 export default Ember.Route.extend(SidePanelRouteMixin, {
+  titleToken: function(model) {
+    let advisorName = model.get('advisor.name');
+
+    return `Unused Advisor: ${advisorName}`;
+  },
+
   model: function(params) {
     return this.store.find('unusedAdvisor', params.unused_advisor_id).then((unusedAdvisor) => {
       let advisorEmails = unusedAdvisor.get('advisor.emails').join(',');
 
       return Ember.RSVP.hash({
-        unusedAdvisor: unusedAdvisor,
-        emails: this.store.find('email', { emails: advisorEmails})
+        emails: this.store.find('email', { emails: advisorEmails}),
+        projectHistory: this.store.find('projectHistory', { advisor_id: unusedAdvisor.get('advisor.id') }),
+        unusedAdvisor: unusedAdvisor
       });
     });
   },
@@ -16,5 +23,6 @@ export default Ember.Route.extend(SidePanelRouteMixin, {
   setupController: function(controller, models) {
     controller.set('model', models.unusedAdvisor);
     controller.set('emails', models.emails);
+    controller.set('projectHistory', models.projectHistory);
   }
 });