+++ app/routes/whiteboards/index.js
@@ -1,7 +1,9 @@
 import Ember from 'ember';
 
 export default Ember.Route.extend({
+  currentUser: Ember.inject.service(),
+
   beforeModel: function() {
-    this.transitionTo('whiteboards.whiteboard', `team-${this.controllerFor('currentUser').get('model.teamId')}`);
+    this.transitionTo('whiteboards.whiteboard', `team-${this.get('currentUser.teamId')}`);
   }
 });