+++ app/routes/whiteboards/whiteboard/project.js
@@ -1,4 +1,8 @@
 import Ember from 'ember';
 import SidePanelRouteMixin from 'ember-cli-paint/mixins/side-panel-route';
 
-export default Ember.Route.extend(SidePanelRouteMixin);
+export default Ember.Route.extend(SidePanelRouteMixin, {
+  titleToken: function(model) {
+    return model.get('name');
+  }
+});