+++ tests/integration/dashboard/interactions-to-schedule-side-panel-test.js
@@ -63,7 +63,8 @@ QUnit.module("Interactions To Schedule Side Panel", {
            "index": 3,
            "created_at": "2015-01-23T21:01:33.615+00:00",
            "angle_ids": [40380],
-           "analyst_1_id": 6565389
+           "analyst_1_id": 6565389,
+           "default_interaction_type": "summarised_call"
         }
       ],
       "interactions": [
@@ -154,11 +155,12 @@ test("Schedule interaction makes an API request and displays a notification", fu
       "project_id": "32522",
       "requested_at": "2015-02-18T10:00:00.000Z",
       "scheduled_call_time": moment().utc().startOf('week').add(1, 'day').add(7, 'hours').toISOString(),
-      "speak": false,
+      "speak": true,
       "speak_phone_number": null,
       "speak_code": null,
       "advisor_phone_number": advisorPhoneNumber,
-      "advisor_phone_country_code": '1'
+      "advisor_phone_country_code": '1',
+      "used": false
     }
   }, response: {
     "interactions": []
@@ -167,6 +169,10 @@ test("Schedule interaction makes an API request and displays a notification", fu
 
   visit('/dashboard');
   click('.interactions-to-schedule article:first');
+
+  andThen(function() {
+    assert.equal($('.ember-select[name=interactionType]').val(), 'summarised_call');
+  });
 
   // Select time slot from calendar
   // Monday 7 AM
@@ -192,6 +198,18 @@ test("Schedule interaction makes an API request and displays a notification", fu
 
     var message = $('.messenger .messenger-message-inner').first().text().trim();
     assert.equal(message, "An interaction between Johnny Advisor and Bob Client has been scheduled.");
+  });
+});
+
+test("Schedule interaction shows errors if validation fails", function(assert) {
+  visit('/dashboard');
+  click('.interactions-to-schedule article:first');
+  click("button:contains('Schedule Interaction')");
+
+  andThen(function() {
+    var label = $('label[for=formattedScheduledCallTime]');
+    var message = label.siblings('.error').text().trim();
+    assert.equal(message, "can't be blank");
   });
 });
 
@@ -225,7 +243,7 @@ test("Cancel interaction returns to dashboard and removes interaction from the w
   });
 
   click('.interactions-to-schedule article:first');
-  click("a:contains('Cancel Interaction')");
+  click("button:contains('Cancel Interaction')");
   click("button:contains('Confirm')");
 
   andThen(function() {
@@ -243,7 +261,10 @@ test("Cancel interaction returns to dashboard and removes interaction from the w
 });
 
 test("Cancel Interaction Failure", function(assert) {
-  defineFixture('DELETE', '/interests/1', { status: 500 });
+  defineFixture('DELETE', '/interests/1', {
+    params: { withdraw_from_compliance: 'false' },
+    status: 500
+  });
 
   visit('/dashboard');
 
@@ -253,7 +274,7 @@ test("Cancel Interaction Failure", function(assert) {
   });
 
   click('.interactions-to-schedule article:first');
-  click("a:contains('Cancel Interaction')");
+  click("button:contains('Cancel Interaction')");
   click("button:contains('Confirm')");
 
   andThen(function() {