+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Phoenix Changelog
 
-### Release 0.0.2 (June 21, 2015)
+### Release 0.0.3 (June 8, 2015)
+
+- [#315 BUGFIX](https://github.com/alphasights/phoenix/pull/315) Fix form submission errors not showing up correctly
+- [#312 BUGFIX](https://github.com/alphasights/phoenix/pull/312) Fix quick jump urls not opening the right pages in Pistachio
+- [#310 BUGFIX](https://github.com/alphasights/phoenix/pull/310) Fix schedule interaction form not working correctly after rescheduling
+- [#308 SECURITY](https://github.com/alphasights/phoenix/pull/308) Enable HSTS headers on Divshot
+- [#301 BUGFIX](https://github.com/alphasights/phoenix/pull/301) Fix widget header and empty view copy bugs
+- [#300 BUGFIX](https://github.com/alphasights/phoenix/pull/300) Fix Team Members filter width
+- [#299](https://github.com/alphasights/phoenix/pull/299) Add basic version of Unused Advisors sidepanel
+- [#296](https://github.com/alphasights/phoenix/pull/296) Cancel Interaction from side panel
+
+### Release 0.0.2 (June 1, 2015)
 
 - [#297](https://github.com/alphasights/phoenix/pull/297) Add Speak numbers to interaction details
 - [#294](https://github.com/alphasights/phoenix/pull/294) Added stats to the projects on the Projects page
-- [BUGFIX #293](https://github.com/alphasights/phoenix/pull/293) Fix checklist item status not displaying the right color
+- [#293 BUGFIX](https://github.com/alphasights/phoenix/pull/293) Fix checklist item status not displaying the right color
 - [#291](https://github.com/alphasights/phoenix/pull/291) Improved interaction details design
 - [#286](https://github.com/alphasights/phoenix/pull/286) Add team switcher to the Projects page